Update charts.py
I still need to do some formatting(colors, profit, etc). I still need to do a groupby sum for the customers and profits because right now it is doubling some names in the top 10 I believe. I had some issues with it for some reason but will try and resolve tomorrow.